+++ packages/core/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact core module, newest chan
 
 ### Added
 
-- `core/util` function `safeChildMap` to safely iterate over Reach `children`
+- `core/util` function `mapAndFilterChildren` to safely iterate over Reach `children`
 
 ## [3.3.0-alpha.13] - 2020-06-22